adapters/rest-py/pyproject.toml
[project]
name = "anip-rest-adapter"
version = "0.1.0"
description = "Generic REST/OpenAPI adapter for any ANIP-compliant service"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
anip-rest-adapter = "anip_rest_adapter.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
